add implementations for 02-bldg/CITY-CODE
