+++ engine/runtime/action-plateau-processor/src/plateau6/mapping.rs
@@ -4,6 +4,7 @@ use once_cell::sync::Lazy;
 use reearth_flow_runtime::node::{NodeKind, ProcessorFactory};
 
 use super::profile::PLATEAU6;
+use crate::common::building_usage_attribute_validator::BuildingUsageAttributeValidatorFactory;
 use crate::common::domain_of_definition_validator::DomainOfDefinitionValidatorFactory;
 use crate::common::missing_attribute_detector::MissingAttributeDetectorFactory;
 use crate::common::object_list_extractor::ObjectListExtractorFactory;
@@ -15,6 +16,7 @@ pub(crate) static ACTION_FACTORY_MAPPINGS: Lazy<HashMap<String, NodeKind>> = Laz
         Box::new(DomainOfDefinitionValidatorFactory::new(&PLATEAU6)),
         Box::new(ObjectListExtractorFactory::new(&PLATEAU6)),
         Box::new(MissingAttributeDetectorFactory::new(&PLATEAU6)),
+        Box::new(BuildingUsageAttributeValidatorFactory::new(&PLATEAU6)),
     ];
     factories
         .into_iter()
